Build fixes
svn path=/trunk/; revision=1452

+++ pkcs11/gck/gck-file-store.c
@@ -135,8 +135,8 @@ read_all_bytes (int fd, guchar *buf, gsize len)
 			if (errno == EAGAIN && errno == EINTR)
 				continue;
 			if (res < 0 || len != all)
-				g_warning ("couldn't read %lu bytes from store file: %s", all, 
-				           g_strerror (errno));
+				g_warning ("couldn't read %u bytes from store file: %s", 
+				           (guint)all, g_strerror (errno));
 			return FALSE;
 		} else  {
 			len -= res;
@@ -160,8 +160,8 @@ write_all_bytes (int fd, const guchar *buf, gsize len)
 		if (res <= 0) {
 			if (errno == EAGAIN && errno == EINTR)
 				continue;
-			g_warning ("couldn't write %lu bytes to store file: %s", all, 
-			           res < 0 ? g_strerror (errno) : "");
+			g_warning ("couldn't write %u bytes to store file: %s", 
+			           (guint)all, res < 0 ? g_strerror (errno) : "");
 			return FALSE;
 		} else  {
 			len -= res;
